--- src/rag/Dockerfile
@@ -1,26 +1,0 @@
-FROM python:3.10-slim
-
-WORKDIR /app
-
-# Install system dependencies
-RUN apt-get update && \
-    apt-get install -y --no-install-recommends \
-    build-essential \
-    gcc \
-    && apt-get clean \
-    && rm -rf /var/lib/apt/lists/*
-
-# Copy requirements file
-COPY requirements.txt .
-
-# Install Python dependencies
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Copy source code
-COPY . .
-
-# Expose port for FastAPI
-EXPOSE 8000
-
-# Command to run the application
-CMD ["uvicorn", "api.fast_api:app", "--host", "0.0.0.0", "--port", "8000"]
