--- src/rag/chroma_db/vector_store.py
@@ -1,1 +1,0 @@
-# Chroma DB component
