+++ src/rag/core/event_specifics/event_finder.py
@@ -1,0 +1,1 @@
+# Event Specifics component
