+++ src/rag/core/prompt_enricher/enricher.py
@@ -1,0 +1,1 @@
+# Prompt Enricher component
