+++ src/rag/docker-compose.yml
@@ -1,38 +1,17 @@
 version: '3.8'
 
 services:
-  rag-api:
-    build: .
-    container_name: rag-api
-    ports:
-      - "8000:8000"
-    volumes:
-      - ./:/app
-    environment:
-      - PYTHONPATH=/app
-      - CHROMA_DB_PATH=/app/data/chroma
-    depends_on:
-      - chroma-db
-    networks:
-      - rag-network
-
-  chroma-db:
-    image: chromadb/chroma:latest
+  chroma:
+    image: ghcr.io/chroma-core/chroma:latest
     container_name: chroma-db
     volumes:
-      - chroma-data:/chroma/data
+      - chroma-data:/chroma/chroma
     ports:
-      - "8001:8000"
-    networks:
-      - rag-network
+      - "8080:8080"
     environment:
-      - ALLOW_RESET=true
-      - CHROMA_SERVER_AUTH_PROVIDER=chromadb.auth.token.TokenAuthenticationProvider
-      - CHROMA_SERVER_AUTH_CREDENTIALS_FILE=/chroma/data/auth.json
-
-networks:
-  rag-network:
-    driver: bridge
+      - CHROMA_HOST=localhost
+      - CHROMA_PORT=8080
+    restart: unless-stopped
 
 volumes:
   chroma-data:
