--- src/rag/event_specifics/event_finder.py
@@ -1,1 +1,0 @@
-# Event Specifics component
