--- src/rag/prompt_enricher/enricher.py
@@ -1,1 +1,0 @@
-# Prompt Enricher component
